Remove legacy core module
