+++ src/app/game/core/index.js
@@ -1,0 +1,38 @@
+export {
+  canPlayCard,
+  playCreature,
+  prepareSpell,
+  handlePassive,
+  activateCreatureAbility,
+  beginTurn,
+  advancePhase,
+  endTurn,
+  startGame,
+  describeGameState,
+  assignBlockerToAttacker,
+  canSelectBlocker,
+  confirmAttackers,
+  describePhase,
+  describePhaseDetailed,
+  finalizeCurrentRequirement,
+  handlePlayerTargetSelection,
+  handleTargetSelection,
+  isAttackingCreature,
+  isTargetableCreature,
+  isTargetablePlayer,
+  prepareBlocks,
+  resolveCombat,
+  scheduleAIPendingConfirmation,
+  scheduleAIPendingResolution,
+  selectBlocker,
+  selectTargetForPending,
+  skipCombat,
+  startCombatStage,
+  toggleAttacker,
+} from './flow.js';
+
+export { drawCards, spendMana, removeFromHand, sortHand, createPlayer } from './players.js';
+export { computeRequirements, buildEffectRequirements, describeRequirement } from './requirements.js';
+export { resolveEffects } from './effects.js';
+export { confirmPendingAction, cancelPendingAction } from './pending.js';
+export { checkForWinner } from './runtime.js';
